refactor: Clarify collection step with Contraption Actors (#9325)
With the expectation that the various method signatures do describe what they do, the "dropItem"
was exceptionally hard to doubt and gone past it several months before come to the conclusion that
maybe there are some side effects implemented where the signature may not suggest.

This does not do any functional change, but improves the readability as now the five places where
this "dropItem" was used, there is a clear indication that the item collection does happen, and
suggests that the context needed to be passed not for position only, but for the contraption
storage as well.